Rename subprocess module to avoid clash with stdlib
Same problem as with #1426 and logging
