+++ setup.cfg
@@ -33,7 +33,7 @@ files =
     uvicorn/server.py,
     uvicorn/__init__.py,
     uvicorn/__main__.py,
-    uvicorn/subprocess.py,
+    uvicorn/_subprocess.py,
     uvicorn/protocols/http/flow_control.py,
     uvicorn/supervisors/statreload.py,
     uvicorn/supervisors/multiprocess.py,
