+++ uvicorn/supervisors/multiprocess.py
@@ -10,7 +10,7 @@ from typing import Callable, List, Optional
 import click
 
 from uvicorn.config import Config
-from uvicorn.subprocess import get_subprocess
+from uvicorn._subprocess import get_subprocess
 
 HANDLED_SIGNALS = (
     signal.SIGINT,  # Unix signal 2. Sent by Ctrl+C.
